login server

+++ server/common/actortype/actor_type.go
@@ -36,7 +36,7 @@ func WorldName(id int32) ActorId {
 func GatewayName(id int32) ActorId {
 	return fmt.Sprintf("%v_%v_Actor", GateWay_Actor, id)
 }
-func ChatName(id uint16) ActorId {
+func ChatName(id int32) ActorId {
 	return fmt.Sprintf("%v_%v_Actor", Chat_Actor, id)
 }
